Refactor inline parsers
